backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
beautifulsoup4>=4.12.3
requests>=2.31.0
easyocr>=1.7.1
numpy>=1.26.3
pillow>=10.2.0
python-dotenv>=1.0.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
razorpay>=1.3.0
phonenumbers>=8.13.0
